+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+* Transferred sources to `src` directory
+* Removed `dispose` method from `PDFDocument` class (Replaced by `close` method) 
+* Added more comments in code
+
 ## 1.0.1
 
 * Fixed support capability dart v2.0 (non 2.1) by add import `dart:core` in files with use `Future`